Made the UI responsive and added table numbers to the orders
