+++ waiter-ui/src/App.js
@@ -1,18 +1,15 @@
 import React, { useState } from "react";
 import { Counter } from "./features/counter/Counter";
-import "./css/App.scss";
 import OrderList from "./components/OrderList";
-import Footer from "./components/Footer";
-import Header from "./components/Header";
-import MainContent from "./components/MainContent";
-
+import "bootstrap/dist/css/bootstrap.min.css";
+// import Footer from "./components/Footer";
+// import Header from "./components/Header";
+// import MainContent from "./components/MainContent";
 
 function App() {
   return (
     <div className="container">
-      <Header></Header>
-      <MainContent></MainContent>
-      <Footer></Footer>
+      <OrderList className="list"></OrderList>
     </div>
   );
 }
