--- waiter-ui/src/components/Footer.js
@@ -1,15 +1,0 @@
-import React from "react";
-
-function footer() {
-  return (
-    <div className="footer">
-      <p>Copyright @2020</p>
-      <div className="social">
-        <i className="fa fa-facebook"></i>
-        <i className="fa fa-instagram"></i>
-      </div>
-    </div>
-  );
-}
-
-export default footer;
