--- waiter-ui/src/components/useOrder.js
@@ -1,30 +1,0 @@
-import { useState } from "react";
-import Button from "react-bootstrap/Button";
-import Modal from "react-bootstrap/Modal";
-//import "bootstrap/dist/css/bootstrap.min.css";
-
-const useOrder = (Boolean) => {
-  const [show, setShow] = useState(Boolean);
-  const handleClose = () => setShow(false);
-
-  return (
-    <>
-      <Modal show={show} onHide={handleClose}>
-        <Modal.Header closeButton>
-          <Modal.Title>Modal heading</Modal.Title>
-        </Modal.Header>
-        <Modal.Body>Woohoo, you're reading this text in a modal!</Modal.Body>
-        <Modal.Footer>
-          <Button variant="secondary" onClick={handleClose}>
-            Close
-          </Button>
-          <Button variant="primary" onClick={handleClose}>
-            Save Changes
-          </Button>
-        </Modal.Footer>
-      </Modal>
-    </>
-  );
-};
-
-export default useOrder;
