--- waiter-ui/src/css/App.scss
@@ -1,175 +1,0 @@
-$round: "Varela Round";
-$cta: #fe8033;
-$bg: #272b34;
-$light_bg: #2b3039;
-$text: #2d343e;
-$white: #fff;
-
-.dFlex {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.container {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  height: 100%;
-  padding: 30px 50px 20px 50px;
-  box-sizing: border-box;
-  background-color: $bg;
-}
-
-body {
-  width: 100%;
-  height: 100%;
-  position: relative;
-  font-family: $round;
-}
-
-nav {
-  @extend .dFlex;
-  justify-content: space-around;
-  width: 100%;
-  color: $white;
-  .logo {
-    font-size: 30px;
-    text-transform: uppercase;
-    color: rgba($white, 0.7);
-  }
-  ul {
-    @extend .dFlex;
-    margin: 0;
-    padding: 0;
-    list-style: none;
-    li {
-      font-size: 14px;
-      margin-left: 15px;
-      padding: 10px;
-      box-sizing: border-box;
-      color: rgba($white, 0.7);
-      cursor: pointer;
-    }
-  }
-
-  .search {
-    color: rgba($white, 0.7);
-    .fa {
-      cursor: pointer;
-      + .fa {
-        margin-left: 15px;
-      }
-    }
-  }
-
-  .main_content {
-    @extend .dFlex;
-    position: relative;
-    width: 70%;
-    height: 100%;
-    margin: 50px auto;
-    padding: 50px 0 0;
-    box-sizing: border-box;
-    justify-content: space-between;
-    flex-wrap: wrap;
-    h3 {
-      position: absolute;
-      left: 0;
-      top: 0;
-      color: $text;
-      font-size: 10.5vw;
-    }
-    .card {
-      width: 30%;
-      border-radius: 15px;
-      @extend .dFlex;
-      flex-direction: column;
-      position: relative;
-      transform: scale(0.8);
-      transition: all 0.3 ease-in-out;
-      margin-bottom: 50px;
-    }
-    .card_img {
-      margin-top: -5vw;
-      img {
-        width: 200px;
-      }
-    }
-    .card_header {
-      @extend .dFlex;
-      flex-direction: column;
-      color: $white;
-      margin: 25px 0;
-      h2 {
-        font-size: 18px;
-      }
-      p {
-        font-size: 14px;
-        text-align: center;
-        color: rgba($white, 0.3);
-        margin: 8px 0;
-      }
-      .price {
-        font-size: 25px;
-        color: $cta;
-        span {
-          font-size: 15px;
-          display: inline-block;
-          vertical-align: top;
-        }
-      }
-    }
-
-    .card:hover {
-      background-color: $light_bg;
-      transform: scale(1.1);
-      cursor: pointer;
-    }
-
-    .card:hover {
-      .btn {
-        background-color: $cta;
-        color: $white;
-      }
-    }
-  }
-}
-.btn {
-  width: 130px;
-  height: 35px;
-  font-size: 14px;
-  border-radius: 35px;
-  background-color: $text;
-  color: rgba($white, 0.5);
-  @extend .dFlex;
-  cursor: pointer;
-  position: relative;
-  transition: background-color 0.3s ease;
-  overflow: hidden;
-}
-
-.footer {
-  @extend .dFlex;
-  width: 100%;
-  justify-content: space-between;
-  margin-top: auto;
-  color: rgba($white, 0.3);
-  font-size: 12px;
-  .social {
-    @extend .dFlex;
-    .fa {
-      width: 50px;
-      height: 50px;
-      border-radius: 50px;
-      border: 1px solid rgba($white, 0.3);
-      @extend .dFlex;
-      font-size: 14px;
-      color: rgba($white, 0.3);
-      cursor: pointer;
-      + .fa {
-        margin-left: 15px;
-      }
-    }
-  }
-}
